feat: node-cron added for scheduled daily challenge
Added node-cron for daily challenge scheduling and modified how setDailychallenge function is set
