+++ backend/package.json
@@ -4,6 +4,7 @@
     "bcrypt": "^6.0.0",
     "cors": "^2.8.5",
     "express": "^5.1.0",
+    "node-cron": "^4.2.1",
     "prisma": "^6.13.0"
   },
   "devDependencies": {
