+++ backend/src/modules/dailygame/dailyRoutes.ts
@@ -5,4 +5,3 @@ export const dailyRouter = express.Router();
 
 // Backend route to get the daily challenge
 dailyRouter.get("/daily", dailyControllers.getDailyChallenge);
-dailyRouter.post("/setdaily", dailyControllers.setDailyChallenge);
